Move tests for log in dedicated directory
